@@ -1,12 +1,15 @@
 package top.theillusivec4.comforts.client;
 
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.client.rendereregistry.v1.BlockEntityRendererRegistry;
+import net.fabricmc.fabric.api.client.rendereregistry.v1.EntityModelLayerRegistry;
 import net.fabricmc.fabric.api.event.client.ClientSpriteRegistryCallback;
-import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
 import net.minecraft.client.render.RenderLayer;
+import net.minecraft.client.render.entity.model.EntityModelLayer;
 import net.minecraft.screen.PlayerScreenHandler;
 import net.minecraft.util.DyeColor;
+import net.minecraft.util.Identifier;
 import top.theillusivec4.comforts.client.renderer.HammockBlockEntityRenderer;
 import top.theillusivec4.comforts.client.renderer.SleepingBagBlockEntityRenderer;
 import top.theillusivec4.comforts.common.ComfortsMod;
@@ -17,8 +20,28 @@ import top.theillusivec4.somnus.api.client.SleepRenderEvents;
 
 public class ComfortsClientMod implements ClientModInitializer {
 
+  public static final EntityModelLayer SLEEPING_BAG_HEAD =
+      new EntityModelLayer(new Identifier(ComfortsMod.MOD_ID, "sleeping_bag_head"), "main");
+  public static final EntityModelLayer SLEEPING_BAG_FOOT =
+      new EntityModelLayer(new Identifier(ComfortsMod.MOD_ID, "sleeping_bag_foot"), "main");
+  public static final EntityModelLayer HAMMOCK_HEAD =
+      new EntityModelLayer(new Identifier(ComfortsMod.MOD_ID, "hammock_head"), "main");
+  public static final EntityModelLayer HAMMOCK_FOOT =
+      new EntityModelLayer(new Identifier(ComfortsMod.MOD_ID, "hammock_foot"), "main");
+
   @Override
   public void onInitializeClient() {
+    EntityModelLayerRegistry.registerModelLayer(SLEEPING_BAG_HEAD,
+        SleepingBagBlockEntityRenderer::getHeadTexturedModelData);
+    EntityModelLayerRegistry.registerModelLayer(
+        SLEEPING_BAG_FOOT,
+        SleepingBagBlockEntityRenderer::getFootTexturedModelData);
+    EntityModelLayerRegistry.registerModelLayer(
+        HAMMOCK_HEAD,
+        HammockBlockEntityRenderer::getHeadTexturedModelData);
+    EntityModelLayerRegistry.registerModelLayer(
+        HAMMOCK_FOOT,
+        HammockBlockEntityRenderer::getFootTexturedModelData);
     BlockEntityRendererRegistry.INSTANCE
         .register(ComfortsRegistry.SLEEPING_BAG_BE, SleepingBagBlockEntityRenderer::new);
     BlockEntityRendererRegistry.INSTANCE
@@ -32,8 +55,8 @@ public class ComfortsClientMod implements ClientModInitializer {
             registry.register(ComfortsMod.id("entity/sleeping_bag/" + color.getName()));
           }
         });
-    ClientSidePacketRegistry.INSTANCE
-        .register(ComfortsNetwork.SYNC_AUTOSLEEP, ComfortsNetwork::readAutoSleep);
+    ClientPlayNetworking
+        .registerGlobalReceiver(ComfortsNetwork.SYNC_AUTOSLEEP, ComfortsNetwork::readAutoSleep);
     SleepRenderEvents.PLAYER_VERTICAL_TRANSLATION
         .register(ComfortsClientEvents::getSleepTranslation);
   }
