@@ -4,7 +4,7 @@ import dev.onyxstudios.cca.api.v3.component.ComponentKey;
 import dev.onyxstudios.cca.api.v3.component.ComponentRegistry;
 import dev.onyxstudios.cca.api.v3.entity.EntityComponentFactoryRegistry;
 import dev.onyxstudios.cca.api.v3.entity.EntityComponentInitializer;
-import nerdhub.cardinal.components.api.util.RespawnCopyStrategy;
+import dev.onyxstudios.cca.api.v3.entity.RespawnCopyStrategy;
 import top.theillusivec4.comforts.common.component.PlayerSleepTracker;
 import top.theillusivec4.comforts.common.component.SleepTrackerComponent;
 
