@@ -1,14 +1,21 @@
 package top.theillusivec4.comforts.common.integration;
 
-import io.github.prospector.modmenu.api.ConfigScreenFactory;
-import io.github.prospector.modmenu.api.ModMenuApi;
-import me.sargunvohra.mcmods.autoconfig1u.AutoConfig;
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import me.shedaniel.autoconfig.AutoConfig;
+import top.theillusivec4.comforts.common.ComfortsMod;
+import top.theillusivec4.comforts.common.config.AutoConfigPlugin;
 import top.theillusivec4.comforts.common.config.ComfortsConfigData;
 
 public class ComfortsModMenu implements ModMenuApi {
 
   @Override
   public ConfigScreenFactory<?> getModConfigScreenFactory() {
-    return screen -> AutoConfig.getConfigScreen(ComfortsConfigData.class, screen).get();
+
+    if (ComfortsMod.isConfigLoaded) {
+      return AutoConfigPlugin::getConfigScreen;
+    } else {
+      return screen -> null;
+    }
   }
 }
